fix(pi): the reserved name is not the whole question [no-changelog]
The shortcut answers whether anything of kendex's is under the name pi
reserved, and the pass was reading that as whether the installation is in
good order. They are different questions and only the first is about that
directory: an older install whose old layout is already gone, whose
registration at the new path somebody moved by hand, took the shortcut
and had the fresh registration written beside theirs — the hook firing
twice, and the pass recording the move as finished on the way past.

The shortcut now asks the one question it was skipping, from the same
place the answer comes from when the old layout is still on disk, so the
two cannot say different things. And nothing an installation is being
held over is recorded as finished, filtered at the one door every answer
leaves by rather than at each of them: a hold is a thing for the person
to fix, and a completion written down while one stands cannot be taken
back.

No changelog entry: what this restores is the promise already written
there — a registration you move at the new path is never doubled by a
second one beside it — which had been true only while the old layout was
still present.
